backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.0
Pillow==10.4.0
python-multipart==0.0.9
requests==2.31.0
openai==1.40.6 